// graphql/generated/graphql.ts
export type BannerQuery = { __typename?: 'Query', banner?: { __typename?: 'Banner', name: string, contents?: Array<{ __typename?: 'ComponentCoreBannerContent', id: string, name: string, sub_heading: string, heading: string, background_image?: { __typename?: 'UploadFile', alternativeText?: string | null, url: string } | null, cta_button?: { __typename?: 'ComponentCoreButton', title: string, id: string, href: string } | null } | null> | null } | null };